first commit
This project is a simple keylogger developed for integration with Discord. The keylogger captures the keystrokes pressed by the user and sends the information to a Discord server, along with details about the window in which the event occurred. Make sure to add your webhook. Please leave suggestions for improvement

For educational purposes only.